requirements.txt
# ---- Data Extraction ----
warcio
requests
beautifulsoup4
lxml
tqdm  # For progress bars on long extracts

# ---- Data Processing & Database ----
pandas      # Used only for loading, not transformation
sqlalchemy
psycopg2-binary

# ---- Transformation (dbt) ----
dbt-core
dbt-postgres

# ---- Entity Matching (LLM) ----
# dbt will handle 95% of matching.
# These are for the final 5% LLM-based match.
anthropic
openai
google-generativeai # Added Gemini as an option

# ---- Data Quality ----
# great-expectations (This is also complex!
# dbt's built-in 'dbt test' is 90% of the value
# for 10% of the effort. I'd cut GE too.)
# For now, let's keep dbt's tests.

# ---- Utilities ----
pyyaml
python-dotenv